requirements.txt
langchain
openai
PyPDF2
faiss-cpu
tiktoken
streamlit
python-dotenv
huggingface_hub
pandas
streamlit_chat